refactor: Extract Select.jsx
