+++ frontend/src/lobby/GameOptions.jsx
@@ -3,8 +3,8 @@ import PropTypes from "prop-types";
 
 import _ from "utils/utils";
 import App from "../app";
-import Checkbox from "../components/Checkbox"
-import { Select } from "../utils";
+import Checkbox from "../components/Checkbox";
+import Select from "../components/Select";
 
 import Set from "./Set";
 import CubeList from "./CubeList";
